Refactor Max BPM renderer into modular package
