+++ src/heart/display/renderers/max_bpm_screen/__init__.py
@@ -1,0 +1,4 @@
+from heart.display.renderers.max_bpm_screen.renderer import \
+    AvatarBpmRenderer as AvatarBpmRenderer
+from heart.display.renderers.max_bpm_screen.renderer import \
+    MaxBpmScreen as MaxBpmScreen
